% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_lags}
\alias{create_lags}
\title{Create lagged variables matrix}
\usage{
create_lags(x, p)
}
\arguments{
\item{x}{Time series vector}

\item{p}{Number of lags}
}
\value{
Matrix of lagged variables
}
\description{
Create lagged variables matrix
}
\keyword{internal}
