% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_suite.R
\name{interpret_results}
\alias{interpret_results}
\title{Interpret test results}
\usage{
interpret_results(results, has_drift)
}
\arguments{
\item{results}{List of test results}

\item{has_drift}{Logical indicating if drift tests were used}
}
\value{
Character string with interpretation
}
\description{
Interpret test results
}
\keyword{internal}
