% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_suite.R
\name{print.kmtest}
\alias{print.kmtest}
\title{Print method for kmtest objects}
\usage{
\method{print}{kmtest}(x, ...)
}
\arguments{
\item{x}{An object of class "kmtest"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for kmtest objects
}
