% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_suite.R
\name{print.kmtest_suite}
\alias{print.kmtest_suite}
\title{Print method for kmtest_suite objects}
\usage{
\method{print}{kmtest_suite}(x, ...)
}
\arguments{
\item{x}{An object of class "kmtest_suite"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for kmtest_suite objects
}
