% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_lag_order}
\alias{select_lag_order}
\title{Select optimal lag order using information criteria}
\usage{
select_lag_order(x, max_p, criterion = "AIC")
}
\arguments{
\item{x}{Time series vector}

\item{max_p}{Maximum lag order to consider}

\item{criterion}{Information criterion: "AIC" or "SIC"}
}
\value{
Optimal lag order
}
\description{
Select optimal lag order using information criteria
}
\keyword{internal}
