% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_suite.R
\name{summary.kmtest}
\alias{summary.kmtest}
\title{Summary method for kmtest objects}
\usage{
\method{summary}{kmtest}(object, ...)
}
\arguments{
\item{object}{An object of class "kmtest"}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame with test summary
}
\description{
Summary method for kmtest objects
}
