% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_suite.R
\name{summary.kmtest_suite}
\alias{summary.kmtest_suite}
\title{Summary method for kmtest_suite objects}
\usage{
\method{summary}{kmtest_suite}(object, ...)
}
\arguments{
\item{object}{An object of class "kmtest_suite"}

\item{...}{Additional arguments (not used)}
}
\value{
The input object invisibly
}
\description{
Summary method for kmtest_suite objects
}
